# openapi/api.yaml
openapi: 3.0.3
info:
  title: RevU API
  description: RevU API for businesses and customers.
  version: 1.0.2
servers:
  - url: 'http://34.209.63.8:5000/api/v1'
    description: Production
  - url: 'http://localhost:5000/api/v1'
    description: Development
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      summary: Create an order
      tags:
        - orders
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
            examples:
              order:
                $ref: '#/components/examples/order'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get orders list
      tags:
        - orders
      operationId: getOrders
      parameters:
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/toDate'
        - name: appointmentId
          description: Limit results to specific appointment id
          in: query
          schema:
            $ref: '#/components/schemas/IntegerId'
        - name: paymentStatus
          description: Limit results to specific payment status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: status
          description: Limit results to specific status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - $ref: '#/components/parameters/orderIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated orders list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete orders
      tags:
        - orders
      operationId: deleteOrders
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Orders deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /orders/{id}:
    get:
      summary: Get an order
      tags:
        - orders
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/orderInclude'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      summary: Update an order
      tags:
        - orders
      operationId: updateOrder
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateBody'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /orders/transactions:
    post:
      summary: Create a transaction
      tags:
        - transactions
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
            examples:
              transaction:
                $ref: '#/components/examples/transaction'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get transactions list
      tags:
        - transactions
      operationId: getTransactions
      parameters:
        - $ref: '#/components/parameters/orderId'
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/toDate'
        - name: isReverse
          description: Limit results to reverse transactions
          in: query
          schema:
            type: boolean
        - name: status
          description: Limit results to specific status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - $ref: '#/components/parameters/paginationInclude'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated transactions list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete transactions
      tags:
        - transactions
      operationId: deleteTransactions
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Transactions deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /orders/transactions/{id}:
    get:
      summary: Get a transaction
      tags:
        - transactions
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      summary: Update a transaction
      tags:
        - transactions
      operationId: updateTransaction
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateBody'
            examples:
              transaction:
                $ref: '#/components/examples/transaction'
      responses:
        '200':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  parameters:
    id:
      name: id
      description: Resource id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    deleteIds:
      name: ids
      description: List of ids to delete
      in: query
      required: true
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    fromDate:
      name: fromDate
      description: Limit results after date
      in: query
      schema:
        $ref: '#/components/schemas/DateTime'
    toDate:
      name: toDate
      description: Limit results before date
      in: query
      schema:
        $ref: '#/components/schemas/DateTime'
...
    orderId:
      name: orderId
      description: Limit results to specific order id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    paginationInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
          - _pagination.total
...
    orderInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - items
    orderIncludeWithPagination:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - items
            - _pagination.total
...
    page:
      name: page
      description: Current page no. for pagination
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    perPage:
      name: perPage
      description: No. of records in paginated list
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    '500':
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PositiveIntegerIncludingZero:
      type: integer
      minimum: 0
    IntegerId:
      type: integer
      minimum: 1
    IntegerIds:
      type: array
      items:
        $ref: '#/components/schemas/IntegerId'
      minItems: 1
    ShortText:
      type: string
      minLength: 2
      maxLength: 64
...
    Money:
      type: number
      format: float
      minimum: 0
...
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 compliant date time
...
    TransactionChannel:
      enum:
        - cash
        - digitalWallet
        - upi
        - card
    TransactionStatus:
      enum:
        - pending
        - completed
        - failed
...
    PaymentStatus:
      enum:
        - pending
        - paid
        - refunded
        - failed
    OrderStatus:
      enum:
        - created
        - fulfilling
        - fulfilled
        - rejected
...
    TransactionUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        transactionId:
          $ref: '#/components/schemas/ShortText'
        channel:
          $ref: '#/components/schemas/TransactionChannel'
        amount:
          $ref: '#/components/schemas/Money'
        orderId:
          $ref: '#/components/schemas/IntegerId'
        isReverse:
          type: boolean
        note:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/DateTime'
          readOnly: true
    Transaction:
      allOf:
        - $ref: '#/components/schemas/TransactionUpdateBody'
      required:
        - channel
        - amount
        - orderId
    OrderUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        type:
          $ref: '#/components/schemas/ShortText'
        appointmentId:
          $ref: '#/components/schemas/IntegerId'
        total:
          $ref: '#/components/schemas/Money'
        discount:
          $ref: '#/components/schemas/Money'
        tax:
          $ref: '#/components/schemas/Money'
        note:
          type: string
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        status:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/DateTime'
          readOnly: true
    Order:
      allOf:
        - $ref: '#/components/schemas/OrderUpdateBody'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
      required:
        - appointmentId
        - total
    OrderItemUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        orderId:
          $ref: '#/components/schemas/IntegerId'
        optedServiceId:
          $ref: '#/components/schemas/IntegerId'
        employeeId:
          $ref: '#/components/schemas/IntegerId'
        price:
          $ref: '#/components/schemas/Money'
        note:
          type: string
    OrderItem:
      allOf:
        - $ref: '#/components/schemas/OrderItemUpdateBody'
      required:
        - optedServiceId
        - price
...
    Error:
      type: object
      properties:
        code:
          oneOf:
            - type: number
            - type: string
        message:
          type: string
        data:
          oneOf:
            - type: object
            - type: array
              items: {}
      required:
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
      required:
        - page
        - perPage
    JSONResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: true
    SingleSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
          required:
            - data
    ListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
          required:
            - data
    PaginatedListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ListSuccessResponse'
        - type: object
          properties:
            _pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - _pagination
...
    MultiObjectOperationSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          properties:
            count:
              $ref: '#/components/schemas/PositiveIntegerIncludingZero'
          required:
            - count
      required:
        - success
        - data
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/Error'
          required:
            - error
  examples:
...
    order:
      value:
        appointmentId: 1
        total: 120
        discount: 10
        tax: 10
...
    transaction:
      value:
        transactionId: TR234455321
        channel: cash
        amount: 100
        orderId: 1
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT